+++ src/common/object/jsc/private_store.hpp
@@ -1,0 +1,29 @@
+////////////////////////////////////////////////////////////////////////////
+//
+// Copyright 2021 Realm Inc.
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+// http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+//
+////////////////////////////////////////////////////////////////////////////
+
+#pragma once
+
+template <typename GetterSetter>
+struct PrivateStore {
+    void *accessor_data = nullptr;
+    ObjectObserver *observer = nullptr;
+    IOCollection *collection = nullptr;
+    std::function<void()> finalizer = nullptr;
+    std::unordered_map<std::string, bool> keys;
+    std::unique_ptr<GetterSetter> getter_setter{nullptr};
+};
