+++ src/common/object/jsc/support.hpp
@@ -1,0 +1,8 @@
+//
+// Created by Cesar Valdez on 13/05/2021.
+//
+
+#ifndef REALMJS_SUPPORT_HPP
+#define REALMJS_SUPPORT_HPP
+
+#endif //REALMJS_SUPPORT_HPP
